Move sendBeacon into tracking
