+++ tests/unit/amo/utils/test_index.js
@@ -7,10 +7,9 @@ import {
   getCanonicalURL,
   getPromotedBadgesLinkUrl,
   makeQueryStringWithUTM,
-  sendBeacon,
   stripLangFromAmoUrl,
 } from 'amo/utils';
-import { getFakeConfig, getFakeLogger } from 'tests/unit/helpers';
+import { getFakeConfig } from 'tests/unit/helpers';
 
 describe(__filename, () => {
   describe('getCanonicalURL', () => {
@@ -257,50 +256,6 @@ describe(__filename, () => {
       const urlString = `https://somehost/${validLang}/${validLang}/somepath/`;
       expect(stripLangFromAmoUrl({ _checkInternalURL, urlString })).toEqual(
         `https://somehost/${validLang}/somepath/`,
-      );
-    });
-  });
-
-  describe('sendBeacon', () => {
-    it('should send a becaon if navigator.sendBeacon exists', () => {
-      const urlString = 'https://www.mozilla.org';
-      const _log = getFakeLogger();
-      const _navigator = { sendBeacon: sinon.spy() };
-
-      sendBeacon({ _log, _navigator, urlString });
-      sinon.assert.calledWith(_log.debug, `Sending beacon to ${urlString}`);
-      sinon.assert.calledWith(_navigator.sendBeacon, urlString);
-    });
-
-    it('can include data in a becaon', () => {
-      const urlString = 'https://www.mozilla.org';
-      const data = 'some-data';
-      const _log = getFakeLogger();
-      const _navigator = { sendBeacon: sinon.spy() };
-
-      sendBeacon({ _log, _navigator, urlString, data });
-      sinon.assert.calledWith(_navigator.sendBeacon, urlString, data);
-    });
-
-    it('should not send a becaon if navigator does not exist', () => {
-      const urlString = 'https://www.mozilla.org';
-      const _log = getFakeLogger();
-
-      sendBeacon({ _log, _navigator: null, urlString });
-      sinon.assert.calledWith(
-        _log.warn,
-        'navigator does not exist. Not sending a beacon.',
-      );
-    });
-
-    it('should not send a becaon if navigator.sendBeacon does not exist', () => {
-      const urlString = 'https://www.mozilla.org';
-      const _log = getFakeLogger();
-
-      sendBeacon({ _log, _navigator: { sendBeacon: null }, urlString });
-      sinon.assert.calledWith(
-        _log.warn,
-        'navigator does not exist. Not sending a beacon.',
       );
     });
   });
